+++ androidplot-core/src/test/java/com/androidplot/util/LinkedLayerListOrganizerTest.java
@@ -6,9 +6,11 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
+import java.util.Arrays;
 import java.util.LinkedList;
 
 import static junit.framework.Assert.assertEquals;
+import static org.junit.Assert.fail;
 
 public class LinkedLayerListOrganizerTest {
     @Before
@@ -28,12 +30,66 @@ public class LinkedLayerListOrganizerTest {
 
     @Test
     public void testMoveAbove() throws Exception {
+        Object obj1 = new Object();
+        Object obj2 = new Object();
+        Object obj3 = new Object();
+        LinkedList<Object> list = new LinkedList<>(Arrays.asList(obj1, obj2, obj3));
+        LayerListOrganizer<Object> organizer = new LayerListOrganizer<>(list);
 
+        organizer.moveAbove(obj1, obj2);
+        assertEquals(Arrays.asList(obj2, obj1, obj3), list);
+
+        organizer.moveAbove(obj2, obj3);
+        assertEquals(Arrays.asList(obj1, obj3, obj2), list);
+    }
+
+    @Test
+    public void testMoveAbove_unknownReference_throwsAndLeavesListUnchanged() throws Exception {
+        Object obj1 = new Object();
+        Object obj2 = new Object();
+        Object obj3 = new Object();
+        LinkedList<Object> list = new LinkedList<>(Arrays.asList(obj1, obj2, obj3));
+        LayerListOrganizer<Object> organizer = new LayerListOrganizer<>(list);
+
+        try {
+            organizer.moveAbove(obj1, new Object());
+            fail("expected IllegalArgumentException");
+        } catch (IllegalArgumentException expected) {
+            // ok
+        }
+        assertEquals(Arrays.asList(obj1, obj2, obj3), list);
     }
 
     @Test
     public void testMoveBeneath() throws Exception {
+        Object obj1 = new Object();
+        Object obj2 = new Object();
+        Object obj3 = new Object();
+        LinkedList<Object> list = new LinkedList<>(Arrays.asList(obj1, obj2, obj3));
+        LayerListOrganizer<Object> organizer = new LayerListOrganizer<>(list);
 
+        organizer.moveBeneath(obj3, obj2);
+        assertEquals(Arrays.asList(obj1, obj3, obj2), list);
+
+        organizer.moveBeneath(obj2, obj1);
+        assertEquals(Arrays.asList(obj2, obj1, obj3), list);
+    }
+
+    @Test
+    public void testMoveBeneath_unknownReference_throwsAndLeavesListUnchanged() throws Exception {
+        Object obj1 = new Object();
+        Object obj2 = new Object();
+        Object obj3 = new Object();
+        LinkedList<Object> list = new LinkedList<>(Arrays.asList(obj1, obj2, obj3));
+        LayerListOrganizer<Object> organizer = new LayerListOrganizer<>(list);
+
+        try {
+            organizer.moveBeneath(obj1, new Object());
+            fail("expected IllegalArgumentException");
+        } catch (IllegalArgumentException expected) {
+            // ok
+        }
+        assertEquals(Arrays.asList(obj1, obj2, obj3), list);
     }
 
     @Test